Move createSystemMLSparkSession to AutomatedTestBase
